+++ Sources/EmojiKit/Extensions/View+Selection.swift
@@ -3,7 +3,7 @@
 //  EmojiKit
 //
 //  Created by Daniel Saidi on 2024-06-21.
-//  Copyright © 2024 Daniel Saidi. All rights reserved.
+//  Copyright © 2024-2025 Daniel Saidi. All rights reserved.
 //
 
 import SwiftUI
@@ -16,17 +16,13 @@ extension View {
         cornerRadius: Double = 10
     ) -> some View {
         if #available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *) {
-            self.background(
-                    ContainerRelativeShape()
-                        .fill(selectionBackgroundFillStyle(isSelected: isSelected))
-                        .aspectRatio(1, contentMode: .fill)
-                )
-            .containerShape(.rect(cornerRadius: cornerRadius))
+            self.background(selectionBackgroundShape(isSelected: isSelected))
+                .containerShape(.rect(cornerRadius: cornerRadius))
         } else {
             self
         }
     }
-
+    
     @available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
     func selectionBackgroundFillStyle(
         isSelected: Bool
@@ -39,5 +35,15 @@ extension View {
         #else
             .init(.clear)
         #endif
+    }
+    
+    @ViewBuilder
+    @available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
+    func selectionBackgroundShape(
+        isSelected: Bool
+    ) -> some View {
+        ContainerRelativeShape()
+            .fill(selectionBackgroundFillStyle(isSelected: isSelected))
+            .aspectRatio(1, contentMode: .fill)
     }
 }